requirements.txt
absl-py==0.6.1
asn1crypto==0.24.0
astor==0.7.1
blinker==1.4
bokeh==1.0.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.0.1
contextlib2==0.5.5
cryptography==2.3.1
cycler==0.10.0
fitbit==0.3.0
gast==0.2.0
grpcio==1.12.1
h5py==2.8.0
idna==2.7
jaraco.functools==1.20
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
llvmlite==0.25.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
numba==0.40.0
numpy==1.15.4
oauthlib==2.1.0
olefile==0.46
packaging==18.0
pandas==0.23.4
patsy==0.5.1
Pillow==5.3.0
portend==2.3
protobuf==3.6.1
pycparser==2.19
PyJWT==1.6.4
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
pywin32==223
PyYAML==3.13
repoze.lru==0.7
requests==2.20.0
requests-oauthlib==1.0.0
Routes==2.4.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
simplejson==3.16.0
six==1.11.0
statsmodels==0.9.0
tempora==1.13
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
tornado==5.1.1
urllib3==1.23
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
zc.lockfile==1.3.0